AOV: Merging changes into aov.cpp
