+++ src/integrators/CMakeLists.txt
@@ -1,7 +1,6 @@
 set(MI_PLUGIN_PREFIX "integrators")
 
 add_plugin(aov        aov.cpp)
-add_plugin(aov_ad     aov_ad.cpp)
 add_plugin(depth      depth.cpp)
 add_plugin(direct     direct.cpp)
 add_plugin(moment     moment.cpp)
